move parts in subpackages
