+++ engine/provider/provider.go
@@ -1,0 +1,13 @@
+package provider
+
+import (
+	"context"
+
+	"go.woodpecker-ci.org/woodpecker/v3/woodpecker-go/woodpecker"
+)
+
+type Provider interface {
+	DeployAgent(context.Context, *woodpecker.Agent) error
+	RemoveAgent(context.Context, *woodpecker.Agent) error
+	ListDeployedAgentNames(context.Context) ([]string, error)
+}
